pyproject.toml
[project]
name = "google-drive-listener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.1",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-storage>=3.1.0",
    "llama-index>=0.12.37",
    "llama-index-cloud-sql-pg>=0.2.2",
    "llama-index-readers-google>=0.6.1",
    "llama-index-vector-stores-pinecone>=0.4.5",
    "llama-parse>=0.6.23",
    "marvin>=3.1.0",
]
